+++ cookbooks/vms/recipes/virtualbox.rb
@@ -31,12 +31,14 @@ box  = node[:box]
 vms  = node[:apps][:vms]
 vbox = vms['profiles']['virtualbox']
 
+vbox_users = box[:users].select { |_, usr| not usr[:guest] }.map { |username, _| username }
+
 # Oracle VM VirtualBox
 install_app "virtualbox" do
   profile vbox
 end
 
-box[:users].select { |_, usr| not usr[:guest] }.each do |username, usr|
+vbox_users.each do |username|
   group "vboxusers" do
     members username
     append true
@@ -49,5 +51,37 @@ vms_vbox_extpack "virtualbox" do
   package vbox['package']
   action :nothing
   subscribes :install, resources("package[virtualbox]"), :immediately
+end
+
+# VirtualBox autostart service
+cookbook_file "vbox-defaults" do
+  path "/etc/default/virtualbox"
+  source "/virtualbox/virtualbox"
+  mode 0644
+  backup false
+end
+
+directory "/etc/vbox" do
+  group "vboxusers"
+  mode 01775
+end
+
+template "vbox-autostart-cfg" do
+  path "/etc/vbox/autostart.cfg"
+  source "/virtualbox/autostart.cfg.erb"
+  mode 0644
+  backup false
+  variables(
+    :users => vbox_users
+  )
+end
+
+bash "vbox-autostart-service" do
+  code <<-EOH
+    update-rc.d -f vboxautostart-service remove
+    update-rc.d vboxautostart-service start 25 2 3 4 5 . stop 75 0 1 6 .
+    EOH
+  action :nothing
+  subscribes :run, resources("template[vbox-autostart-cfg]"), :immediately
 end